backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-dramatiq==0.11.2
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dramatiq==1.14.2
drf-yasg==1.21.5
exceptiongroup==1.1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
pluggy==1.0.0
prometheus-client==0.17.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pytest==7.3.1
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.4
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
zope.interface==6.0

APScheduler~=3.10.1